refactor(services/integration): change name of module
