+++ crates/services/integration/src/sink/kodi.rs
@@ -5,7 +5,7 @@ use media_models::{ImportOrExportMediaItem, ImportOrExportMediaItemSeen};
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
-use crate::traita::YankIntegration;
+use crate::utils::YankIntegration;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 struct IntegrationMediaSeen {
